Rename condition_kernel_fatbin.h to graph_do_while_cond_fatbin.h
