+++ scripts/build_condition_kernel_fatbin.sh
@@ -3,7 +3,7 @@
 # Builds the pre-compiled fatbin for the graph_do_while condition kernel.
 # Requires the CUDA toolkit (nvcc) to be installed.
 #
-# The output is a C header (condition_kernel_fatbin.h) containing the fatbin
+# The output is a C header (graph_do_while_cond_fatbin.h) containing the fatbin
 # as a byte array, which is compiled into the quadrants binary. At runtime
 # this lets us load the condition kernel without needing libcudadevrt.a on
 # the user's system.
@@ -41,7 +41,7 @@ echo "Generating C header ..."
 python3 -c "
 import sys
 data = open('$FATBIN', 'rb').read()
-out = open('$OUT_DIR/condition_kernel_fatbin.h', 'w')
+out = open('$OUT_DIR/graph_do_while_cond_fatbin.h', 'w')
 out.write('// Auto-generated by scripts/build_condition_kernel_fatbin.sh\n')
 out.write('// Do not edit manually. Regenerate with that script.\n')
 out.write('//\n')
